+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Send email when the payment is successful
 - Transform `mjml` files (email's template files) to html and plaintext
 - Add email settings and configure `mailcatcher` in docker-compose stack
 - Add Open Badge Factory badge provider